test/mochitest/test_detect_b2g.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1235459
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container">
  <h3 id="offline"></h3>
  <ul id="devices"></ul>
  <input id="userBuild" type="file" value="Upload your own build" />
  <button id="installBtn">Flash Device</button>
  <div class="dialog" id="confirmDialog"><div>
    <section>
      <p>Your device has succesfully installed Firefox OS!</p>
       <button>Ok</button>
    </section>
  </div></div>
</div>
<pre id="test">
  <script type="application/javascript" id="fake-adb"></script>
  <script type="application/javascript" id="code"></script>
  <script type="application/javascript">

  "use strict";

  var fakeB2GDevice, fakeAndroidDevice;

  function populateFakeDevices() {
    return loadFakeDevice().then(() => {
      fakeB2GDevice = new FakeAdbDevice();
      fakeB2GDevice._files = { "/system/b2g/application.ini": "" };

      fakeAndroidDevice = new FakeAdbDevice();
      fakeAndroidDevice._files = { "/system/bin/surface_flinger": "" };
    });
  }

  function test_detect_b2gdevice() {
    return checkDeviceIsB2G(fakeB2GDevice).then(isB2G => {
      is(isB2G, true, "detected a B2G device");
    });
  }

  function test_detect_androiddevice() {
    return checkDeviceIsB2G(fakeAndroidDevice).then(isB2G => {
      is(isB2G, false, "detected an Android device");
    });
  }

  function runTest() {
    loadAboutJs()
      .then(() => populateFakeDevices())
      .then(() => test_detect_b2gdevice())
      .then(() => test_detect_androiddevice())
      .then(() => SimpleTest.finish());
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
